% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexes.R
\docType{data}
\name{complexes}
\alias{complexes}
\title{Data of complexes}
\format{
A tibble with 182 rows and 3 variables:
\describe{
\item{species}{dbl Full species name}
\item{code}{dbl Code}
\item{scale}{dbl Scale of a complex}
}
}
\source{
\url{link}
}
\usage{
complexes
}
\description{
To be added later
}
\keyword{datasets}
