% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eu.R
\docType{data}
\name{data_eu}
\alias{data_eu}
\title{Data of testate amoebae samples from Amesbury at al., 2016}
\format{
A tibble with 1302 rows and 55 variables:
\describe{
\item{person}{}
\item{country}{}
\item{site}{}
\item{sample}{}
\item{lat}{}
\item{long}{}
\item{wtd}{}
\item{ph}{}
}
}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0277379116303742}
}
\usage{
data_eu
}
\description{
Testate amoebae species are pooled in larger groups, environmental data includes WTD and pH variables
}
\keyword{datasets}
